pyproject.toml
[tool.poetry]
name = "salesforce-analyzer"
version = "0.1.0"
description = "A comprehensive tool for analyzing and documenting Salesforce org metadata and automations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"          # For configuration files
click = "^8.1.3"         # For CLI interface
rich = "^13.3.1"         # For beautiful terminal output
requests = "^2.28.2"     # For API interactions
pandas = "^2.0.0"        # For data processing
networkx = "^3.0"        # For graph operations
torch = "^2.0.0"         # For local LLM support
transformers = "^4.27.0" # For LLM models
simple-salesforce = "^1.12.1"  # For Salesforce API interactions
mermaid-py = "^0.1.1"         # For generating Mermaid diagrams
typing-extensions = "^4.5.0"   # For enhanced type hints
python-dotenv = "^1.0.0"       # For environment variable management
accelerate = ">=0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
black = "^23.3.0"        # Code formatter
isort = "^5.12.0"       # Import sorter
mypy = "^1.3.0"         # Static type checker
pylint = "^2.17.0"      # Code linter
pre-commit = "^3.2.0"   # Git hooks
pytest-mock = "^3.10.0" # For mocking in tests
pytest-asyncio = "^0.21.0"
bandit = "^1.7.5"       # Security linter

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
salesforce-analyzer = "src.cli:cli"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "R0903",  # too-few-public-methods
    "C0103",  # invalid-name
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src"
testpaths = [
    "tests",
]